Rename test pacage controller in to ServiceAtClinicController
