+++ api/src/main/java/SpringProject/_Spring/dto/ServiceAtClinicMapper.java
@@ -5,11 +5,11 @@ import SpringProject._Spring.model.ServiceAtClinic;
 import java.util.List;
 
 public class ServiceAtClinicMapper {
-    public static ServiceAtClinicDTO toServiceAtClinicDTO(ServiceAtClinic service){
-        return new ServiceAtClinicDTO(service.getId(), service.getName(), service.getDescription(), service.getPrice());
+    public static ServiceAtClinicResponseDTO toServiceAtClinicDTO(ServiceAtClinic service){
+        return new ServiceAtClinicResponseDTO(service.getId(), service.getName(), service.getDescription(), service.getPrice());
     }
 
-    public static List<ServiceAtClinicDTO> toServiceAtClinicDTO(List<ServiceAtClinic> services){
-        return services.stream().map( s ->  new ServiceAtClinicDTO(s.getId(), s.getName(), s.getDescription(), s.getPrice())).toList();
+    public static List<ServiceAtClinicResponseDTO> toServiceAtClinicDTO(List<ServiceAtClinic> services){
+        return services.stream().map( s ->  new ServiceAtClinicResponseDTO(s.getId(), s.getName(), s.getDescription(), s.getPrice())).toList();
     }
 }
